[REM] *: accounting files not in their module's manifest
*: l10n_cl, l10n_nz, sale_expense

Description of the issue this commit addresses:

[This PR](https://github.com/odoo/odoo/pull/211150) has brought our attention to some files that were in the codebase but
not included in their module's manifest. Therefore they are useless as is and
can either be deleted or need to be put in the manifest.

Desired behavior after this commit is merged:

Unused useless files have been removed from the codebase.
Unused useful files have been added to their module's manifest.

task-4822341

closes odoo/odoo#211816

X-original-commit: 132379aad6e5f7ee1f380f5104309c6471e0c12c
Related: odoo/enterprise#86502
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>