[FIX] website_event: set an opacity to demo's "An unpublished event"
Setting the opacity to the empty string leads to max opacity but is seen
as "None" in the builder because no options corresponds to this value.
And the "None" option is not that, which is confusing. Setting a known
value for the opacity shows the correct choice for the option.

Steps to reproduce:
- In a database with demo data, go to `/event`
- Go to "An unpublished event"
- Open editor
- Bug: "Filter Intensity" shows "None"

Backport of 94cc1cbd2d8ed23fd243515ad215ea91bac46068

task-4367641

closes odoo/odoo#213965

X-original-commit: 36485f468b037aca7caf58ba86a10542379e0d43
Signed-off-by: Soukéina Bojabza (sobo) <sobo@odoo.com>